Added batch support for Azure
